Extract user banning functionality to a concern
